+++ ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: drubio-m <drubio-m@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 13:44:32 by drubio-m          #+#    #+#             */
-/*   Updated: 2022/03/22 12:55:34 by drubio-m         ###   ########.fr       */
+/*   Updated: 2022/03/25 22:54:01 by drubio-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <string.h>
+
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
